# src/main/openapi/action-api.yaml
openapi: 3.1.0
info:
  title: Portable Agent Action API
  version: 1.0.0
  description: Создание, просмотр и подтверждение безопасных действий агента.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://api.portable-agent.dev/action-service
security:
  - bearerAuth: []
paths:
  /api/v1/actions:
    post:
      operationId: proposeAction
      summary: Создать предложение действия
      tags: [Actions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeActionRequest"
      responses:
        "201":
          description: Предложение создано и ожидает policy evaluation или подтверждения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /api/v1/actions/{actionId}:
    get:
      operationId: getAction
      summary: Получить текущее состояние действия
      tags: [Actions]
      parameters:
        - $ref: "#/components/parameters/ActionId"
      responses:
        "200":
          description: Текущее состояние
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/actions/{actionId}/decisions:
    post:
      operationId: decideAction
      summary: Подтвердить или отменить действие
      tags: [Actions]
      parameters:
        - $ref: "#/components/parameters/ActionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionDecisionRequest"
      responses:
        "202":
          description: Решение принято durable workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "409":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ActionId:
      name: actionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: Ошибка в формате RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ProposeActionRequest:
      type: object
      additionalProperties: false
      required: [kind, connector, payload, requestKey]
      properties:
        kind:
          type: string
          enum: [calendar.create_event]
        connector:
          type: string
          enum: [fake-calendar]
        payload:
          type: object
          minProperties: 1
          additionalProperties: true
        requestKey:
          type: string
          minLength: 8
          maxLength: 128
    ActionDecisionRequest:
      type: object
      additionalProperties: false
      required: [decision, payloadHash]
      properties:
        decision:
          type: string
          enum: [CONFIRM, CANCEL]
        payloadHash:
          type: string
          pattern: "^[a-f0-9]{64}$"
    ActionResponse:
      type: object
      additionalProperties: false
      required:
        [
          id,
          status,
          kind,
          connector,
          payload,
          payloadHash,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum:
            [
              PROPOSED,
              AWAITING_APPROVAL,
              APPROVED,
              EXECUTING,
              SUCCEEDED,
              FAILED,
              CANCELLED,
            ]
        kind: { type: string }
        connector: { type: string }
        payload:
          type: object
          additionalProperties: true
          description: Сохранённые данные, которые пользователь подтверждает по payloadHash.
        payloadHash: { type: string, pattern: "^[a-f0-9]{64}$" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }